requirements.txt
Django==1.11.2
Jinja2==2.9.6
psycopg2==2.7.3.1
pytz==2017.2
six==1.10.0
virtualenv==15.1.0
virtualenvwrapper==4.7.2
django-filter==1.0.4
djangorestframework==3.6.4
markdown==2.6.9
coreapi==2.3.1
coreschema==0.0.4
itypes==1.1.0
uritemplate==3.0.0
urllib3==1.22